April 5 2024 tour

+++ database/migrations/2024_04_03_045049_create_locations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('service');
-            $table->double('price', 8, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamps();
         });
     }
